Refactored: Moved ChatInterface to mainwindow.py (Fixes #83)
